requirements.txt
amqp==2.6.1
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
billiard==3.6.3.0
boto3==1.17.33
botocore==1.20.33
cairocffi==1.2.0
CairoSVG==2.5.2
celery==4.4.7
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.6
cssselect2==0.4.1
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.44.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==2.0.7
django-rest-auth==0.9.5
django-storages==1.11.1
django-timezone-field==4.1.2
djangorestframework==3.12.2
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
factory-boy==3.2.0
Faker==6.6.2
gunicorn==20.0.4
html5lib==1.1
idna==2.10
inflection==0.5.1
ipython==7.21.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
kombu==4.6.11
Markdown==3.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.18
psycopg2-binary==2.8.5
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
PyJWT==2.0.1
pyparsing==2.4.7
Pyphen==0.10.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.2.0
pytz==2018.5
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
s3transfer==0.3.6
six==1.14.0
sqlparse==0.4.1
text-unidecode==1.3
tinycss2==1.1.0
traitlets==5.0.5
typing==3.6.4
uritemplate==3.0.1
urllib3==1.26.4
vine==1.3.0
wcwidth==0.2.5
WeasyPrint==52.4
webencodings==0.5.1
